README file
BRAIN TUMOR DETECTION USING MATLAB

Brain Tumor is a fatal disease which cannot be confidently detected without MRI. In the project, it is tried to detect whether patient’s brain has tumor or not from MRI image using MATLAB simulation. To pave the way for morphological operation on MRI image, the image was first  filtered using Anisotropic Diffusion Filter to reduce contrast between consecutive pixels. After that the image was resized and utilizing a threshold value image was converted to a black and white image manually. This primary filter the plausible locations for tumor presence. On this semi processed image morphological operations have been applied and information on solidity and areas of the plausible locations was obtained. A minimum value of both of this characters has been determined from statistical average of different MRI images containing tumor. Then it was used to deliver final detection result.